Adapt to inheritance changes in matplotlib Axes
The matplotlib 'Axes' class no longer inherits from 'SubplotBase'
since https://github.com/matplotlib/matplotlib/pull/23573.

This breaks the '_get_subplot' method in 'sampling_point.py' which
checks if the passed 'axes_obj' is a 'SubplotBase' instance to
distinguish between indexable and non-indexable objects.

To fix this, we now use a try-except block to index into the
'axes_obj', and catch the 'TypeError' that is raised if it is
and 'Axes' (as opposed to an array).