Enable using distributed RST files.
Use a list of paths instead of a single path in the 'rst' member
of 'ContinuousFiberCompositesFiles', to allow passing multiple
RST files.
Add a constructor that can take either a single path or a list of
paths, and converts it to a list if needed.

Rename the 'rst_file' argument to 'rst_files' in the
'ResultDefinition' class, and convert it to a list.

Adapt the following functions to handle multiple RST files:
* get_composite_data_sources
* upload_continuous_fiber_composite_files_to_server
* get_composite_files_from_workbench_result_folder
* get_continuous_fiber_example_files

Add a 'material_support' member to 'CompositeDataSources', since
DPF core cannot (currently) use a distributed DataSources object for
creating the material support. Adapt the 'get_composite_data_sources'
function accordingly. This is a temporary workaround, until DPF core
supports creating a material support from a distributed DataSources.