Update pyvista, and mark it as optional (#304)
Update the PyVista dependency. Since it is only used explicitly
in the docs, mark it as optional.
Add the '[jupyter]' extra to the PyVista dependency, since it is
needed for the docs build since PyVista 0.40.

Remove the explicit dependency on 'vtk', since it is a transitive
dependency that's not directly used.