Move minimum_requirements.txt to separate directory
Dependabot struggles to get the correct lockfile to update when
both 'poetry.lock' and 'minimum_requirements.txt' are located in
the same directory and creates weird PR descriptions. See e.g.
https://github.com/ansys/pydpf-composites/pull/351, where
the actual base version is '0.42.1', but the PR mentions
'0.36.1' which is the version from 'minimum_requirements.txt'.

As a simple workaround, we move the 'minimum_requirements.txt' to
a separate directory.