Add Python 3.13 support, drop 3.9
Remove the upper bound of the supported Python version, since that
makes it tedious to add new versions: PyDPF Composites needs to
be released before we can update PyACP, for example.

Add Python 3.13 to the explicitly supported versions in the
trove classifiers, and add it to the test matrix.

Drop support for Python 3.9.

Run pyupgrade with the '--py310-plus' flag, and remove the now
unused imports.