Remove y-axis margin for sampling point plots (#604)
In the sampling point plots, set the y-axis margin to zero. When
a margin is present, it may be confused with an additional layer
where the data is missing. By removing the margin, the plot bounds
exactly match the data range.

Use plt.show() instead of fig.show() in the rst workflow example,
to fix rendering when directly running the script.