Run old server version tests in separate CI jobs (#606)
Run the tests with container tags other than `latest` on a separate CI job, to reduce
the time for CI to finish.

Rename `image-tag` to `container-tag` everywhere for consistency with the `CONTAINER_TAG` 
option used elsewhere in `ci_cd.yml`.

The older versions are now tested only with Python 3.13, as opposed to testing
all combinations of Python version + server version.